requirements.txt
python-dotenv==1.0.0
pytz==2023.3.post1
selenium==4.14.0
selenium-wire==5.1.0
slack-sdk==3.21.3
